refactor: 객체의 성격 별로 패키지 분리
