+++ app/src/main/java/woowacourse/shopping/data/datasource/CartDataSource.kt
@@ -1,4 +1,6 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.data.datasource
+
+import woowacourse.shopping.data.model.CartItem
 
 interface CartDataSource {
     fun getCartItems(
